+++ src/xrCore/stream_reader.cpp
@@ -53,51 +53,52 @@ void CStreamReader::map						(const u32 &new_offset)
 	m_start_pointer				= m_current_pointer;
 }
 
-void CStreamReader::advance					(int offset)
+void CStreamReader::advance(size_t offset)
 {
-	VERIFY						(m_current_pointer >= m_start_pointer);
-	VERIFY						(u32(m_current_pointer - m_start_pointer) <= m_current_window_size);
+	VERIFY(m_current_pointer >= m_start_pointer);
+	VERIFY(u32(m_current_pointer - m_start_pointer) <= m_current_window_size);
 	int							offset_inside_window = int(m_current_pointer - m_start_pointer);
 	if (offset_inside_window + offset >= (int)m_current_window_size) {
-		remap					(m_current_offset_from_start + offset_inside_window + offset);
+		remap(m_current_offset_from_start + offset_inside_window + offset);
 		return;
 	}
 
 	if (offset_inside_window + offset < 0) {
-		remap					(m_current_offset_from_start + offset_inside_window + offset);
+		remap(m_current_offset_from_start + offset_inside_window + offset);
 		return;
 	}
 
-	m_current_pointer			+= offset;
+	m_current_pointer += offset;
 }
 
-void CStreamReader::r						(void *_buffer, u32 buffer_size)
+void CStreamReader::r(void* _buffer, size_t buffer_size)
 {
-	VERIFY						(m_current_pointer >= m_start_pointer);
-	VERIFY						(u32(m_current_pointer - m_start_pointer) <= m_current_window_size);
+	VERIFY(m_current_pointer >= m_start_pointer);
+	VERIFY(u32(m_current_pointer - m_start_pointer) <= m_current_window_size);
 
-	int							offset_inside_window = int(m_current_pointer - m_start_pointer);
-	if (offset_inside_window + buffer_size < m_current_window_size) {
-		Memory.mem_copy			(_buffer,m_current_pointer,buffer_size);
-		m_current_pointer		+= buffer_size;
+	int offset_inside_window = int(m_current_pointer - m_start_pointer);
+	if (offset_inside_window + buffer_size < m_current_window_size)
+	{
+		Memory.mem_copy(_buffer, m_current_pointer, buffer_size);
+		m_current_pointer += buffer_size;
 		return;
 	}
 
-	u8							*buffer = (u8*)_buffer;
-	u32							elapsed_in_window = m_current_window_size - u32(m_current_pointer - m_start_pointer);
+	u8* buffer = (u8*)_buffer;
+	u32 elapsed_in_window = m_current_window_size - u32(m_current_pointer - m_start_pointer);
 
-	do {
-		Memory.mem_copy			(buffer,m_current_pointer,elapsed_in_window);
-		buffer					+= elapsed_in_window;
-		buffer_size				-= elapsed_in_window;
-		advance					(elapsed_in_window);
+	do
+	{
+		Memory.mem_copy(buffer, m_current_pointer, elapsed_in_window);
+		buffer += elapsed_in_window;
+		buffer_size -= elapsed_in_window;
+		advance(elapsed_in_window);
 
-		elapsed_in_window		= m_current_window_size;
-	}
-	while (m_current_window_size < buffer_size);
+		elapsed_in_window = m_current_window_size;
+	} while (m_current_window_size < buffer_size);
 
-	Memory.mem_copy				(buffer,m_current_pointer,buffer_size);
-	advance						(buffer_size);
+	Memory.mem_copy(buffer, m_current_pointer, buffer_size);
+	advance(buffer_size);
 }
 
 CStreamReader *CStreamReader::open_chunk	(const u32 &chunk_id)
