+++ src/xrCore/stream_reader_inline.h
@@ -45,28 +45,28 @@ IC void CStreamReader::remap(const u32& new_offset)
 	map(new_offset);
 }
 
-IC u32 CStreamReader::elapsed() const
+IC size_t CStreamReader::elapsed() const
 {
 	u32 offset_from_file_start = tell();
 	VERIFY(m_file_size >= offset_from_file_start);
 	return (m_file_size - offset_from_file_start);
 }
 
-IC u32 CStreamReader::length() const
+IC size_t CStreamReader::length() const
 {
 	return (m_file_size);
 }
 
-IC void CStreamReader::seek(int offset)
+IC void CStreamReader::seek(size_t offset)
 {
 	advance(offset - tell());
 }
 
-IC	u32 CStreamReader::tell() const
+IC size_t CStreamReader::tell() const
 {
 	VERIFY(m_current_pointer >= m_start_pointer);
 	VERIFY(u32(m_current_pointer - m_start_pointer) <= m_current_window_size);
-	return			u32(m_current_offset_from_start + (m_current_pointer - m_start_pointer));
+	return u32(m_current_offset_from_start + (m_current_pointer - m_start_pointer));
 }
 
 IC	void CStreamReader::close()
